+++ CODEOWNERS
@@ -219,6 +219,7 @@ models/perf/ @yieldthought @uaydonat
 models/perf/perf_report.py @yieldthought @uaydonat
 models/perf/benchmarking_utils.py @skhorasganiTT @williamlyTT @uaydonat
 models/demos/utils/llm_demo_utils.py @skhorasganiTT @mtairum @uaydonat
+models/experimental/stable_diffusion_xl_bas @mbezuljTT @ppopovic @ipotkonjak-tt @tenstorrent/metalium-developers-convolutions
 
 # docs
 docs/Makefile @tenstorrent/metalium-developers-infra
