+++ models/experimental/stable_diffusion_xl_base/tests/pcc/test_module_tt_geglu.py
@@ -4,7 +4,7 @@
 import torch
 import pytest
 import ttnn
-from models.experimental.stable_diffusion_xl_base_1_0.ttnn_impl.tt_geglu import TtGEGLU
+from models.experimental.stable_diffusion_xl_base.ttnn_impl.tt_geglu import TtGEGLU
 from diffusers import DiffusionPipeline
 from tests.ttnn.utils_for_testing import assert_with_pcc
 from models.utility_functions import torch_random
