+++ models/experimental/stable_diffusion_xl_base/tests/pcc/test_module_tt_upsample2d.py
@@ -4,22 +4,22 @@
 import torch
 import pytest
 import ttnn
-from models.experimental.stable_diffusion_xl_base_1_0.ttnn_impl.tt_downsample2d import TtDownsample2D
+from models.experimental.stable_diffusion_xl_base.ttnn_impl.tt_upsample2d import TtUpsample2D
 from diffusers import DiffusionPipeline
 from tests.ttnn.utils_for_testing import assert_with_pcc
 from models.utility_functions import torch_random
-from models.experimental.stable_diffusion_xl_base_1_0.ttnn_impl.sdxl_utility import (
+from models.experimental.stable_diffusion_xl_base.ttnn_impl.sdxl_utility import (
     to_channel_last_ttnn,
     from_channel_last_ttnn,
 )
 
 
-@pytest.mark.parametrize("input_shape, down_block_id", [((1, 320, 128, 128), 0), ((1, 640, 64, 64), 1)])
-@pytest.mark.parametrize("stride", [(2, 2)])
+@pytest.mark.parametrize("input_shape, up_block_id", [((1, 1280, 32, 32), 0), ((1, 640, 64, 64), 1)])
+@pytest.mark.parametrize("stride", [(1, 1)])
 @pytest.mark.parametrize("padding", [(1, 1)])
 @pytest.mark.parametrize("dilation", [(1, 1)])
 @pytest.mark.parametrize("device_params", [{"l1_small_size": 16384}], indirect=True)
-def test_downsample2d(device, input_shape, down_block_id, stride, padding, dilation, use_program_cache):
+def test_upsample2d(device, input_shape, up_block_id, stride, padding, dilation, use_program_cache):
     pipe = DiffusionPipeline.from_pretrained(
         "stabilityai/stable-diffusion-xl-base-1.0", torch_dtype=torch.float32, use_safetensors=True, variant="fp16"
     )
@@ -27,19 +27,19 @@ def test_downsample2d(device, input_shape, down_block_id, stride, padding, dilat
     unet.eval()
     state_dict = unet.state_dict()
 
-    torch_downsample = unet.down_blocks[down_block_id].downsamplers[0]
+    torch_upsample = unet.up_blocks[up_block_id].upsamplers[0]
     groups = 1
-    tt_downsample = TtDownsample2D(
-        device, state_dict, f"down_blocks.{down_block_id}.downsamplers.0", stride, padding, dilation, groups
+    tt_upsample = TtUpsample2D(
+        device, state_dict, f"up_blocks.{up_block_id}.upsamplers.0", stride, padding, dilation, groups
     )
 
     torch_input_tensor = torch_random(input_shape, -0.1, 0.1, dtype=torch.float32)
-    torch_output_tensor = torch_downsample(torch_input_tensor)
+    torch_output_tensor = torch_upsample(torch_input_tensor)
 
     ttnn_input_tensor = to_channel_last_ttnn(torch_input_tensor, ttnn.bfloat16, device, ttnn.DRAM_MEMORY_CONFIG)
-    ttnn_output_tensor, output_shape = tt_downsample.forward(ttnn_input_tensor, input_shape)
+    ttnn_output_tensor, output_shape = tt_upsample.forward(ttnn_input_tensor)
     output_tensor = from_channel_last_ttnn(
         ttnn_output_tensor, [input_shape[0], output_shape[0], output_shape[1], torch_output_tensor.shape[1]]
     )
 
-    assert_with_pcc(torch_output_tensor, output_tensor, 0.998)
+    assert_with_pcc(torch_output_tensor, output_tensor, 0.996)
