+++ models/experimental/stable_diffusion_xl_base/ttnn_impl/tt_downsample2d.py
