+++ models/experimental/stable_diffusion_xl_base/ttnn_impl/tt_resnetblock2d.py
