--- tests/nightly/single_card/stable_diffusion_xl/test_module_tt_upsample2d.py
@@ -1,1 +1,0 @@
-models/experimental/stable_diffusion_xl_base_1_0/tests/pcc/test_module_tt_upsample2d.py
